moved verify and upload to ui folder
